separate out model to individual user and messages
