+++ Unit-02/01-blueprints/project/messages/views.py
@@ -1,5 +1,6 @@
 from flask import redirect, render_template, request, url_for, Blueprint, flash
-from project.models import Message, User
+from project.messages.models import Message
+from project.users.models import User
 from project.messages.forms import MessageForm, DeleteForm
 from project import db
