+++ Unit-02/01-blueprints/project/users/models.py
@@ -14,14 +14,3 @@ class User(db.Model):
     self.first_name = first_name
     self.last_name = last_name
 
-class Message(db.Model):
-
-  __tablename__ = "messages"
-
-  id = db.Column(db.Integer, primary_key=True)
-  content = db.Column(db.Text)
-  user_id = db.Column(db.Integer, db.ForeignKey('users.id'))
-
-  def __init__(self, content, user_id):
-    self.content = content
-    self.user_id = user_id
